splits non-Interface code out of slicecam_interface in preparation for slicecam acquisition mode. no code changes yet.
